third_party/blink/web_tests/fast/invalid/residual-style-expected.txt

CONSOLE WARNING: A descendant of a <select> does not follow the content model.
CONSOLE WARNING: A descendant of a <select> does not follow the content model.
CONSOLE WARNING: A descendant of a <select> does not follow the content model.
CONSOLE WARNING: A descendant of a <select> does not follow the content model.
CONSOLE WARNING: A descendant of a <select> does not follow the content model.
